created model base class for common methods; created model subclass for each entity
